add more decorator tests
