+++ test/unit/test_decorators.py
@@ -1,117 +1,260 @@
+# Test case for property
+# more tests are in test_descr
+
+import sys
 import unittest
 
-log1 = []
-log2 = []
-cnt = 0
+class PropertyBase(Exception):
+    pass
 
-def decofunc(fn):
-    log1.append(str(fn))
-    return fn
+class PropertyGet(PropertyBase):
+    pass
 
-@decofunc
-def func(x):
-    log1.append(x)
+class PropertySet(PropertyBase):
+    pass
 
-class TestFunctionDecoratorOnFunction(unittest.TestCase):
-    def setup(self):
-        log1 = []
+class PropertyDel(PropertyBase):
+    pass
 
-    def test_function_on_function(self):
-        func('help')
-        self.assertEqual(log1, ['<function func>', 'help'])
+class BaseClass(object):
+    def __init__(self):
+        self._spam = 5
 
-class cdeco(object):
-    def __init__(self, fget=None, fset=None, fdel=None, doc=None):
-        global cnt
-        self.id = cnt
-        cnt += 1
-        log2.append("cdeco.__init__" + str(self.id))
-        log2.append("  " + str(fget))
-        log2.append("  " + str(fset))
-        log2.append("  " + str(fdel))
-        self.fget = fget
-        self.fset = fset
-        self.fdel = fdel
+    @property
+    def spam(self):
+        """BaseClass.getter"""
+        return self._spam
 
-    def __get__(self, obj, loc):
-        log2.append("cdeco.__get__" + str(self.id))
-        return self.fget(obj)
+    @spam.setter
+    def spam(self, value):
+        self._spam = value
 
-    def __set__(self, obj, value):
-        log2.append("cdeco.__set__" + str(self.id))
-        if self.fset is None:
-            raise AttributeError("can't set attribute")
-        self.fset(obj, value)
+    @spam.deleter
+    def spam(self):
+        pass
+        #del self._spam
 
-    def __delete__(self, obj):
-        log2.append("cdeco.__delete__" + str(self.id))
-        if self.fdel is None:
-            raise AttributeError("can't delete attribute")
-        self.fdel(obj)
+class SubClass(BaseClass):
 
-    def getter(self, fset):
-        log2.append("cdeco.getter" + str(self.id))
-        return type(self)(fget, self.fset, self.fdel)
+    @BaseClass.spam.getter
+    def spam(self):
+        """SubClass.getter"""
+        raise PropertyGet(self._spam)
 
-    def setter(self, fset):
-        log2.append("cdeco.setter" + str(self.id))
-        return type(self)(self.fget, fset, self.fdel)
+    @spam.setter
+    def spam(self, value):
+        raise PropertySet(self._spam)
 
-    def deleter(self, fdel):
-        log2.append("cdeco.deleter" + str(self.id))
-        return type(self)(self.fget, self.fset, fdel)
+    @spam.deleter
+    def spam(self):
+        raise PropertyDel(self._spam)
 
-class testclass:
-    def __init__(self, val):
-        self._val = val;
+class PropertyDocBase(object):
+    _spam = 1
+    def _get_spam(self):
+        return self._spam
+    spam = property(_get_spam, doc="spam spam spam")
 
-    @cdeco
-    def val(self):
-        log2.append("testclass.val - getter")
-        return self._val
+class PropertyDocSub(PropertyDocBase):
+    @PropertyDocBase.spam.getter
+    def spam(self):
+        """The decorator does not use this doc string"""
+        return self._spam
 
-    @val.setter
-    def val(self, val):
-        log2.append("testclass.val - setter")
-        self._val = val;
+class PropertySubNewGetter(BaseClass):
+    @BaseClass.spam.getter
+    def spam(self):
+        """new docstring"""
+        return 5
 
-    @val.deleter
-    def val(self):
-        log2.append("testclass.val - deleter")
+class PropertyNewGetter(object):
+    @property
+    def spam(self):
+        """original docstring"""
+        return 1
+    @spam.getter
+    def spam(self):
+        """new docstring"""
+        return 8
 
-class TestDescriptorGetSetOnMethod(unittest.TestCase):
-    def setup(self):
-        log2 = []
+class PropertyTests(unittest.TestCase):
+    def test_property_decorator_baseclass(self):
+        # see #1620
+        base = BaseClass()
+        self.assertEqual(base.spam, 5)
+        self.assertEqual(base._spam, 5)
+        base.spam = 10
+        self.assertEqual(base.spam, 10)
+        self.assertEqual(base._spam, 10)
+        # delattr(base, "spam")
+        # self.assertTrue(not hasattr(base, "spam"))
+        # self.assertTrue(not hasattr(base, "_spam"))
+        base.spam = 20
+        self.assertEqual(base.spam, 20)
+        self.assertEqual(base._spam, 20)
 
-    def test_handmade_descriptor(self):
-        y = testclass(123)
-        log2.append(y.val)
-        y.val = 456
-        log2.append(y.val)
-        self.assertEqual(log2, [
-            'cdeco.__init__0',
-            '  <function val>',
-            '  None',
-            '  None',
-            'cdeco.setter0',
-            'cdeco.__init__1',
-            '  <function val>',
-            '  <function val>',
-            '  None',
-            'cdeco.deleter1',
-            'cdeco.__init__2',
-            '  <function val>',
-            '  <function val>',
-            '  <function val>',
-            'cdeco.__get__2',
-            'testclass.val - getter',
-            123,
-            'cdeco.__set__2',
-            'testclass.val - setter',
-            'cdeco.__get__2',
-            'testclass.val - getter',
-            456])
-        #del y.val del is not yet implemented
+    def test_property_decorator_subclass(self):
+        # see #1620
+        sub = SubClass()
+        self.assertRaises(PropertyGet, getattr, sub, "spam")
+        self.assertRaises(PropertySet, setattr, sub, "spam", None)
+        # self.assertRaises(PropertyDel, delattr, sub, "spam")
 
-if __name__ == "__main__":
+    # def test_property_decorator_subclass_doc(self):
+    #     sub = SubClass()
+    #     self.assertEqual(sub.__class__.spam.__doc__, "SubClass.getter")
+    #
+    # def test_property_decorator_baseclass_doc(self):
+    #     base = BaseClass()
+    #     self.assertEqual(base.__class__.spam.__doc__, "BaseClass.getter")
+
+    # def test_property_decorator_doc(self):
+    #     base = PropertyDocBase()
+    #     sub = PropertyDocSub()
+    #     self.assertEqual(base.__class__.spam.__doc__, "spam spam spam")
+    #     self.assertEqual(sub.__class__.spam.__doc__, "spam spam spam")
+
+    # def test_property_getter_doc_override(self):
+    #     newgettersub = PropertySubNewGetter()
+    #     self.assertEqual(newgettersub.spam, 5)
+    #     self.assertEqual(newgettersub.__class__.spam.__doc__, "new docstring")
+    #     newgetter = PropertyNewGetter()
+    #     self.assertEqual(newgetter.spam, 8)
+    #     self.assertEqual(newgetter.__class__.spam.__doc__, "new docstring")
+
+    # def test_property___isabstractmethod__descriptor(self):
+    #     for val in (True, False, [], [1], '', '1'):
+    #         class C(object):
+    #             def foo(self):
+    #                 pass
+    #             foo.__isabstractmethod__ = val
+    #             foo = property(foo)
+    #         self.assertIs(C.foo.__isabstractmethod__, bool(val))
+    #
+    #     # check that the property's __isabstractmethod__ descriptor does the
+    #     # right thing when presented with a value that fails truth testing:
+    #     class NotBool(object):
+    #         def __bool__(self):
+    #             raise ValueError()
+    #         __len__ = __bool__
+    #     with self.assertRaises(ValueError):
+    #         class C(object):
+    #             def foo(self):
+    #                 pass
+    #             foo.__isabstractmethod__ = NotBool()
+    #             foo = property(foo)
+    #         C.foo.__isabstractmethod__
+
+    def test_property_builtin_doc_writable(self):
+        p = property(doc='basic')
+        self.assertEqual(p.__doc__, 'basic')
+        p.__doc__ = 'extended'
+        self.assertEqual(p.__doc__, 'extended')
+
+    # def test_property_decorator_doc_writable(self):
+    #     class PropertyWritableDoc(object):
+    #
+    #         @property
+    #         def spam(self):
+    #             """Eggs"""
+    #             return "eggs"
+    #
+    #     sub = PropertyWritableDoc()
+    #     self.assertEqual(sub.__class__.spam.__doc__, 'Eggs')
+    #     sub.__class__.spam.__doc__ = 'Spam'
+    #     self.assertEqual(sub.__class__.spam.__doc__, 'Spam')
+
+# Issue 5890: subclasses of property do not preserve method __doc__ strings
+class PropertySub(property):
+    """This is a subclass of property"""
+
+class PropertySubSlots(property):
+    """This is a subclass of property that defines __slots__"""
+    __slots__ = ()
+
+class Foo(object):
+    def __init__(self): self._spam = 1
+    @PropertySub
+    def spam(self):
+        """spam wrapped in property subclass"""
+        return self._spam
+    @spam.setter
+    def spam(self, value):
+        """this docstring is ignored"""
+        self._spam = value
+
+class FooSub(Foo):
+    @Foo.spam.setter
+    def spam(self, value):
+        """another ignored docstring"""
+        self._spam = 'eggs'
+
+class PropertySubclassTests(unittest.TestCase):
+
+    # def test_slots_docstring_copy_exception(self):
+    #     try:
+    #         class Foo(object):
+    #             @PropertySubSlots
+    #             def spam(self):
+    #                 """Trying to copy this docstring will raise an exception"""
+    #                 return 1
+    #     except AttributeError:
+    #         pass
+    #     else:
+    #         raise Exception("AttributeError not raised")
+    #
+    # def test_docstring_copy(self):
+    #     class Foo(object):
+    #         @PropertySub
+    #         def spam(self):
+    #             """spam wrapped in property subclass"""
+    #             return 1
+    #     self.assertEqual(
+    #         Foo.spam.__doc__,
+    #         "spam wrapped in property subclass")
+
+    def test_property_setter_copies_getter_docstring(self):
+        foo = Foo()
+        self.assertEqual(foo.spam, 1)
+        foo.spam = 2
+        self.assertEqual(foo.spam, 2)
+        # self.assertEqual(
+        #     Foo.spam.__doc__,
+        #     "spam wrapped in property subclass")
+
+        foosub = FooSub()
+        self.assertEqual(foosub.spam, 1)
+        foosub.spam = 7
+        self.assertEqual(foosub.spam, 'eggs')
+        # self.assertEqual(
+        #     FooSub.spam.__doc__,
+        #     "spam wrapped in property subclass")
+
+    # def test_property_new_getter_new_docstring(self):
+    #
+    #     class Foo(object):
+    #         @PropertySub
+    #         def spam(self):
+    #             """a docstring"""
+    #             return 1
+    #         @spam.getter
+    #         def spam(self):
+    #             """a new docstring"""
+    #             return 2
+    #     self.assertEqual(Foo.spam.__doc__, "a new docstring")
+    #     class FooBase(object):
+    #         @PropertySub
+    #         def spam(self):
+    #             """a docstring"""
+    #             return 1
+    #     class Foo2(FooBase):
+    #         @FooBase.spam.getter
+    #         def spam(self):
+    #             """a new docstring"""
+    #             return 2
+    #     self.assertEqual(Foo.spam.__doc__, "a new docstring")
+
+
+
+if __name__ == '__main__':
     unittest.main()
